+++ package.json
@@ -29,6 +29,7 @@
     "normalize.css": "4.2.0",
     "passport": "0.3.2",
     "passport-facebook": "2.1.1",
+    "passport-google-oauth2": "^0.1.6",
     "pretty-error": "2.0.0",
     "react": "15.3.1",
     "react-dom": "15.3.1",
